replace old transaction-level reorg handling (via receipt queries) with block-level reorg detection by comparing block hashes to find the divergence point; add `PaymentReorged`, `PaymentConfirmed`, `PaymentFailed` chain events, and `required_confirmations` to `PaymentDetected`; add `Pending` status (dropped transaction) to `payments` schema; fix `safe_lag` binding type in `postgres.rs` (line 242); delete duplicate `idx_invoices_address` index migration

+++ necko3-database/src/backends/postgres.rs
@@ -239,7 +239,7 @@ impl ChainStore for PostgresAdapter {
             .bind(&chain_config.logo_url)
             .bind(chain_config.watch_addresses.iter()
                 .collect::<Vec<_>>())
-            .bind(chain_config.safe_lag)
+            .bind(chain_config.safe_lag as i16)
             .execute(&self.pool)
             .await?;
